+++ android/app/src/main/kotlin/deckers/thibault/aves/channel/calls/DebugHandler.kt
@@ -112,9 +112,7 @@ class DebugHandler(private val context: Context) : MethodCallHandler {
                     isVideo(mimeType) -> ContentUris.withAppendedId(MediaStore.Video.Media.EXTERNAL_CONTENT_URI, id)
                     else -> uri
                 }
-                if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.Q) {
-                    contentUri = MediaStore.setRequireOriginal(contentUri)
-                }
+                contentUri = StorageUtils.getOriginalUri(contentUri)
             }
         }