+++ android/app/src/main/kotlin/deckers/thibault/aves/channel/calls/StorageHandler.kt
@@ -29,6 +29,7 @@ class StorageHandler(private val context: Context) : MethodCallHandler {
             "getInaccessibleDirectories" -> safe(call, result, ::getInaccessibleDirectories)
             "getRestrictedDirectories" -> safe(call, result, ::getRestrictedDirectories)
             "revokeDirectoryAccess" -> safe(call, result, ::revokeDirectoryAccess)
+            "deleteEmptyDirectories" -> safe(call, result, ::deleteEmptyDirectories)
             "scanFile" -> GlobalScope.launch(Dispatchers.IO) { safe(call, result, ::scanFile) }
             else -> result.notImplemented()
         }
@@ -134,6 +135,27 @@ class StorageHandler(private val context: Context) : MethodCallHandler {
 
         val success = PermissionManager.revokeDirectoryAccess(context, path)
         result.success(success)
+    }
+
+    private fun deleteEmptyDirectories(call: MethodCall, result: MethodChannel.Result) {
+        val dirPaths = call.argument<List<String>>("dirPaths")
+        if (dirPaths == null) {
+            result.error("deleteEmptyDirectories-args", "failed because of missing arguments", null)
+            return
+        }
+
+        var deleted = 0
+        dirPaths.forEach {
+            try {
+                val dir = File(it)
+                if (dir.isDirectory && dir.listFiles()?.isEmpty() == true && dir.delete()) {
+                    deleted++
+                }
+            } catch (e: SecurityException) {
+                // ignore
+            }
+        }
+        result.success(deleted)
     }
 
     private fun scanFile(call: MethodCall, result: MethodChannel.Result) {