+++ lib/widgets/viewer/entry_horizontal_pager.dart
@@ -3,10 +3,10 @@ import 'package:aves/model/multipage.dart';
 import 'package:aves/model/source/collection_lens.dart';
 import 'package:aves/widgets/common/magnifier/pan/gesture_detector_scope.dart';
 import 'package:aves/widgets/common/magnifier/pan/scroll_physics.dart';
+import 'package:aves/widgets/common/video/video.dart';
 import 'package:aves/widgets/viewer/multipage.dart';
 import 'package:aves/widgets/viewer/visual/entry_page_view.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_ijkplayer/flutter_ijkplayer.dart';
 import 'package:provider/provider.dart';
 import 'package:tuple/tuple.dart';
 
@@ -15,7 +15,7 @@ class MultiEntryScroller extends StatefulWidget {
   final PageController pageController;
   final ValueChanged<int> onPageChanged;
   final VoidCallback onTap;
-  final List<Tuple2<String, IjkMediaController>> videoControllers;
+  final List<Tuple2<String, AvesVideoController>> videoControllers;
   final List<Tuple2<String, MultiPageController>> multiPageControllers;
   final void Function(String uri) onViewDisposed;
 
@@ -89,7 +89,7 @@ class _MultiEntryScrollerState extends State<MultiEntryScroller> with AutomaticK
           mainEntry: entry,
           page: page,
           viewportSize: mqSize,
-          onTap: (_) => widget.onTap?.call(),
+          onTap: widget.onTap == null ? null : (_) => widget.onTap(),
           videoControllers: widget.videoControllers,
           onDisposed: () => widget.onViewDisposed?.call(entry.uri),
         );
@@ -108,7 +108,7 @@ class _MultiEntryScrollerState extends State<MultiEntryScroller> with AutomaticK
 class SingleEntryScroller extends StatefulWidget {
   final AvesEntry entry;
   final VoidCallback onTap;
-  final List<Tuple2<String, IjkMediaController>> videoControllers;
+  final List<Tuple2<String, AvesVideoController>> videoControllers;
   final List<Tuple2<String, MultiPageController>> multiPageControllers;
 
   const SingleEntryScroller({
@@ -163,7 +163,7 @@ class _SingleEntryScrollerState extends State<SingleEntryScroller> with Automati
           mainEntry: entry,
           page: page,
           viewportSize: mqSize,
-          onTap: (_) => widget.onTap?.call(),
+          onTap: widget.onTap == null ? null : (_) => widget.onTap(),
           videoControllers: widget.videoControllers,
         );
       },