+++ test/model/filters_test.dart
@@ -6,7 +6,6 @@ import 'package:aves/model/filters/mime.dart';
 import 'package:aves/model/filters/query.dart';
 import 'package:aves/model/filters/tag.dart';
 import 'package:aves/model/filters/type.dart';
-import 'package:aves/ref/mime_types.dart';
 import 'package:test/test.dart';
 
 void main() {
@@ -16,16 +15,16 @@ void main() {
     final album = AlbumFilter('path/to/album', 'album');
     expect(album, jsonRoundTrip(album));
 
-    final fav = FavouriteFilter();
+    const fav = FavouriteFilter.instance;
     expect(fav, jsonRoundTrip(fav));
 
     final location = LocationFilter(LocationLevel.country, 'France${LocationFilter.locationSeparator}FR');
     expect(location, jsonRoundTrip(location));
 
-    final type = TypeFilter(TypeFilter.sphericalVideo);
+    final type = TypeFilter.sphericalVideo;
     expect(type, jsonRoundTrip(type));
 
-    final mime = MimeFilter(MimeTypes.anyVideo);
+    final mime = MimeFilter.video;
     expect(mime, jsonRoundTrip(mime));
 
     final query = QueryFilter('some query');